Reshape test script
