+++ testsuite/python/CMakeLists.txt
@@ -264,7 +264,8 @@ python_test(FILE long_range_actors.py MAX_NUM_PROC 4 GPU_SLOTS 2)
 python_test(FILE tabulated.py MAX_NUM_PROC 2)
 python_test(FILE particle_slice.py MAX_NUM_PROC 4)
 python_test(FILE rigid_bond.py MAX_NUM_PROC 4)
-python_test(FILE rigid_bond_virial.py MAX_NUM_PROC 4 LABELS long)
+python_test(FILE rigid_bond_virial.py MAX_NUM_PROC 4)
+python_test(FILE rigid_bond_virial_stats.py MAX_NUM_PROC 4 LABELS long)
 python_test(FILE rotation_per_particle.py MAX_NUM_PROC 4)
 python_test(FILE rotational_inertia.py MAX_NUM_PROC 2)
 python_test(FILE rotational-diffusion-aniso.py MAX_NUM_PROC 1 LABELS long)
