+++ testsuite/python/rigid_bond_virial.py
@@ -77,7 +77,7 @@ class RigidBondVirialTest(ut.TestCase):
     def _virial_unequal_masses(self, set_integrator):
         m1 = 1.0
         m2 = 2.0
-        d = 1.0
+        d = 1.0  # distance between particle
         v = 1.0  # velocity
         V = self.system.volume()
 
@@ -106,50 +106,6 @@ class RigidBondVirialTest(ut.TestCase):
         """SE: constraint virial with m1!=m2 matches centripetal theory."""
         self._virial_unequal_masses(
             self.system.integrator.set_symplectic_euler)
-
-    #  Langevin consistency: mean of (P_bond + P_kin) = kT/V,
-    #  std matches analytic fluctuation formula.
-    def _virial_consistency(self, set_integrator, noise_prefactor):
-        kT = 1.0
-        gamma = 1.0
-        mass = 1.
-        V = self.system.volume()
-        dt = self.system.time_step
-        self.system.thermostat.set_langevin(kT=kT, gamma=gamma, seed=42)
-        set_integrator()
-
-        std_theory = ((6 * kT**2 + noise_prefactor * gamma *
-                      mass * kT / dt) / (9 * V**2))**0.5
-
-        self._make_dimer(v=0.0)
-
-        self.system.integrator.run(1000)   # equilibrate
-        n_loop = 2000
-        n_steps = 100
-        virial = []
-        for _ in range(n_loop):
-            self.system.integrator.run(n_steps)
-            v_p = np.trace(
-                self.system.analysis.pressure_tensor()['bonded']) / 3.
-            v_k = np.trace(self.system.analysis.pressure_tensor()[
-                           'kinetic']) / 3.
-            virial.append(v_p + v_k)
-
-        rigid_p = np.mean(virial)
-        rigid_std = np.std(virial)
-        self.assertAlmostEqual(
-            rigid_p, 1. / V, delta=2. * std_theory / n_loop**0.5)
-        self.assertAlmostEqual(rigid_std, std_theory, delta=0.02 * std_theory)
-
-    def test_virial_consistency_vv(self):
-        """VV+Langevin: rigid bond virial satisfies equipartition and fluctuation formula."""
-        self._virial_consistency(self.system.integrator.set_vv, 1)
-
-    def test_virial_consistency_se(self):
-        """SE+Langevin: rigid bond virial satisfies equipartition and fluctuation formula."""
-        self._virial_consistency(
-            self.system.integrator.set_symplectic_euler, 4)
-
 
 if __name__ == "__main__":
     ut.main()
